HARP-8067: Rename @here/harp-omv-datasource to @here/harp-vectortile-datasource
This change renames the module `@here/harp-omv-datasource` to
`@here/harp-vectortile-datasource` and it adds a compat
module `@here/harp-omv-datasource` to keep backward source
compatibility.
