+++ .prettierignore
@@ -1,4 +1,4 @@
-@here/harp-omv-datasource/lib/proto/vector_tile.d.ts
+@here/harp-vectortile-datasource/lib/proto/vector_tile.d.ts
 @here/harp-datasource-protocol/theme.schema.json
 @here/harp-map-theme/resources/fonts/
 @here/harp-text-canvas/resources/fonts/
