+++ @here/harp-geojson-datasource/package.json
@@ -24,7 +24,7 @@
     },
     "license": "Apache-2.0",
     "dependencies": {
-        "@here/harp-omv-datasource": "^0.18.0",
+        "@here/harp-vectortile-datasource": "^0.18.0",
         "@here/harp-utils": "^0.18.0"
     },
     "devDependencies": {
