+++ @here/harp-omv-datasource/index.ts
@@ -12,10 +12,4 @@
  * @packageDocumentation
  */
 
-export * from "./lib/OmvDataFilter";
-export * from "./lib/OmvDataSource";
-export * from "./lib/OmvDecoderDefs";
-export * from "./lib/OmvTile";
-export * from "./lib/OmvDebugLabelsTile";
-export * from "./lib/OmvRestClient";
-export * from "./lib/GeoJsonDataProvider";
+export * from "@here/harp-vectortile-datasource";
