+++ @here/harp-vectortile-datasource/.gitignore
@@ -1,0 +1,11 @@
+.vscode
+.idea
+node_modules
+dist
+*.d.ts
+*.js
+*.js.map
+!./lib/typings/earcut.d.ts
+!./lib/proto/vector_tile.js
+!./lib/proto/vector_tile.d.ts
+*.tgz
