+++ @here/harp-vectortile-datasource/.npmignore
@@ -1,0 +1,10 @@
+*.ts
+!*.d.ts
+test/
+tsconfig.json
+.gitignore
+.gitreview
+*.proto
+*.tgz
+*.map
+*.tsbuildinfo
