+++ @here/harp-vectortile-datasource/.yarnrc
@@ -1,0 +1,2 @@
+version-git-message "@here/harp-vectortile-datasource %s"
+version-tag-prefix "@here/harp-vectortile-datasource@"
