+++ @here/harp-vectortile-datasource/index.ts
@@ -1,0 +1,21 @@
+/*
+ * Copyright (C) 2017-2020 HERE Europe B.V.
+ * Licensed under Apache 2.0, see full license in LICENSE
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * HERE OMV Data source.
+ *
+ * @remarks
+ *
+ * @packageDocumentation
+ */
+
+export * from "./lib/OmvDataFilter";
+export * from "./lib/OmvDataSource";
+export * from "./lib/OmvDecoderDefs";
+export * from "./lib/OmvTile";
+export * from "./lib/OmvDebugLabelsTile";
+export * from "./lib/OmvRestClient";
+export * from "./lib/GeoJsonDataProvider";
