+++ @here/harp-vectortile-datasource/lib/DecodeInfo.ts
