+++ @here/harp-vectortile-datasource/lib/GeoJsonDataProvider.ts
