+++ @here/harp-vectortile-datasource/lib/IGeometryProcessor.ts
