+++ @here/harp-vectortile-datasource/lib/OmvData.ts
