+++ @here/harp-vectortile-datasource/lib/OmvDataFilter.ts
