+++ @here/harp-vectortile-datasource/lib/OmvDebugLabelsTile.ts
