+++ @here/harp-vectortile-datasource/lib/OmvDecodedTileEmitter.ts
