+++ @here/harp-vectortile-datasource/lib/OmvDecoder.ts
