+++ @here/harp-vectortile-datasource/lib/OmvDecoderDefs.ts
