+++ @here/harp-vectortile-datasource/lib/OmvPoliticalViewFeatureModifier.ts
