+++ @here/harp-vectortile-datasource/lib/OmvRestClient.ts
