+++ @here/harp-vectortile-datasource/lib/OmvTile.ts
