+++ @here/harp-vectortile-datasource/lib/OmvTiler.ts
