+++ @here/harp-vectortile-datasource/lib/OmvTomTomFeatureModifier.ts
