+++ @here/harp-vectortile-datasource/lib/OmvUtils.ts
