+++ @here/harp-vectortile-datasource/lib/Ring.ts
