+++ @here/harp-vectortile-datasource/lib/StyleSetDataFilter.ts
