+++ @here/harp-vectortile-datasource/lib/TiledGeoJsonAdapter.ts
