+++ @here/harp-vectortile-datasource/lib/VTJsonDataAdapter.ts
