+++ @here/harp-vectortile-datasource/lib/proto/vector_tile.proto
