+++ @here/harp-vectortile-datasource/package.json
@@ -1,11 +1,11 @@
 {
-    "name": "@here/harp-omv-datasource",
+    "name": "@here/harp-vectortile-datasource",
     "version": "0.18.0",
-    "description": "HERE OMV Data source",
+    "description": "HERE Vector Tile Data Source",
     "main": "index.js",
     "typings": "index",
     "worker": {
-        "name": "@here/harp-omv-datasource-worker",
+        "name": "@here/harp-vectortile-datasource-worker",
         "main": "index-worker.js"
     },
     "directories": {
@@ -13,13 +13,14 @@
     },
     "scripts": {
         "build": "tsc --build $EXTRA_TSC_ARGS",
-        "test": "cross-env mocha --require source-map-support/register $EXTRA_MOCHA_ARGS ./test/*.js",
+        "test": "cross-env mocha --require source-map-support/register $EXTRA_MOCHA_ARGS ./test/OmvTest.js",
+        "gen-decoder": "pbjs -t static-module -w commonjs --no-encode --no-verify --no-create lib/proto/vector_tile.proto -o lib/proto/vector_tile.js && pbts lib/proto/vector_tile.js -o lib/proto/vector_tile.d.ts",
         "prepare": "cross-env tsc --build $EXTRA_TSC_ARGS"
     },
     "repository": {
         "type": "git",
         "url": "https://github.com/heremaps/harp.gl.git",
-        "directory": "@here/harp-omv-datasource"
+        "directory": "@here/harp-vectortile-datasource"
     },
     "author": {
         "name": "HERE Europe B.V.",
@@ -27,7 +28,20 @@
     },
     "license": "Apache-2.0",
     "dependencies": {
-        "@here/harp-vectortile-datasource": "^0.18.0"
+        "@here/harp-datasource-protocol": "^0.18.0",
+        "@here/harp-geometry": "^0.18.0",
+        "@here/harp-geoutils": "^0.18.0",
+        "@here/harp-lines": "^0.18.0",
+        "@here/harp-lrucache": "^0.18.0",
+        "@here/harp-mapview": "^0.18.0",
+        "@here/harp-mapview-decoder": "^0.18.0",
+        "@here/harp-materials": "^0.18.0",
+        "@here/harp-text-canvas": "^0.18.0",
+        "@here/harp-transfer-manager": "^0.18.0",
+        "@here/harp-utils": "^0.18.0",
+        "earcut": "^2.2.2",
+        "long": "^4.0.0",
+        "protobufjs": "^6.9.0"
     },
     "devDependencies": {
         "@here/harp-fetch": "^0.18.0",
