+++ @here/harp-vectortile-datasource/test/MapViewPickingTest.ts
