+++ @here/harp-vectortile-datasource/test/OmvDecodedTileEmitterTest.ts
