+++ @here/harp-vectortile-datasource/test/OmvTomTomFeatureModifierTest.ts
