+++ @here/harp-vectortile-datasource/test/RingTest.ts
