+++ @here/harp-vectortile-datasource/test/resources/geoJsonData.ts
