+++ test/performance/package.json
@@ -13,6 +13,7 @@
         "@here/harp-lines": "^0.18.0",
         "@here/harp-mapview": "^0.18.0",
         "@here/harp-omv-datasource": "^0.18.0",
+        "@here/harp-vectortile-datasource": "^0.18.0",
         "@here/harp-test-utils": "^0.18.0",
         "@types/chai": "^4.2.11",
         "chai": "^4.2.0",
