+++ test/rendering/package.json
@@ -14,7 +14,7 @@
         "@here/harp-geoutils": "^0.18.0",
         "@here/harp-mapview": "^0.18.0",
         "@here/harp-mapview-decoder": "^0.18.0",
-        "@here/harp-omv-datasource": "^0.18.0",
+        "@here/harp-vectortile-datasource": "^0.18.0",
         "@here/harp-test-utils": "^0.18.0",
         "@here/harp-utils": "^0.18.0",
         "@types/chai": "^4.2.11",
